fix(imq): correct FOCUS_SCORE overflow + vet 4 image-quality features vs analytic oracle
FOCUS_SCORE and LOCAL_FOCUS_SCORE were dominated by an unsigned-integer overflow: focus_score.cpp's
laplacian() multiplied PixIntens (unsigned) pixels by the negative Laplacian kernel weights, so e.g.
-4 wrapped to ~4.29e9 and FOCUS_SCORE reached 2.81e18. Cast the pixel to double before the multiply
-> FOCUS_SCORE 12.109, LOCAL_FOCUS_SCORE 2.681.

Vet 4 of the 6 imq features against an independent numpy/scipy reimplementation of their documented
algorithms (tests/vetting/oracles/gen_imq_analytic.py; reproduces every value exactly):
  - FOCUS_SCORE       = variance of the Laplacian magnitude (scipy convolution) = 12.109
  - LOCAL_FOCUS_SCORE = same on the top-left h/2 x w/2 tile / 4 = 2.681
  - MIN_SATURATION    = fraction of pixels at the image minimum = 18/96 = 0.1875
  - MAX_SATURATION    = fraction of pixels at the image maximum = 16/96 = 0.16667
imq 0/6 -> 4/6.

The other two stay regression with documented reasons (oracle_coverage.csv notes):
  - POWER_SPECTRUM_SLOPE: 0.0 on this fixture (rps() early-returns when min(h,w)/8 < 3), and its
    radial binning uses the FFT value as the radius index instead of the frequency radius -- a real
    bug; needs a rewrite + a >=24 px fixture.
  - SHARPNESS: correct (all-double DOM sharpness port, no bug); needs a reference-DOM oracle.

gtest 717/717; pytest 72 passed (7 arrow fails are the tiff-only local build; the pytest imq tests
check the ImageQuality class runs, not feature values). Coverage 604 -> 608 / 758.

+++ src/nyx/features/focus_score.cpp
@@ -254,7 +254,10 @@ void FocusScoreFeature::laplacian(const std::vector<PixIntens>& image, std::vect
 
                     if(ii >= 0 && ii < m_image && jj >= 0 && jj < n_image &&
                        ikFlip >= 0 && jkFlip >=0 && ikFlip < m_kernel && jkFlip < n_kernel){
-                        out[i* n_image + j] += image[ii * n_image + jj] * kernel[ikFlip * n_kernel + jkFlip];
+                        // Cast the pixel to double before multiplying: image is PixIntens (unsigned),
+                        // so the negative kernel weights (e.g. -4) were converted to a huge unsigned
+                        // value, wrapping the Laplacian (focus score reached ~1e18 from the overflow).
+                        out[i* n_image + j] += (double)image[ii * n_image + jj] * kernel[ikFlip * n_kernel + jkFlip];
                     }
                 }
             }
